requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions==1.17.0
pandas==2.1.0
requests==2.31.0
lxml==4.9.3
azure-storage-blob==12.17.0
pyjwt==2.8.0
gitpython==3.1.40
PyGithub==2.1.1
papermill==2.5.0
nbconvert==7.11.0
seaborn==0.13.0
matplotlib==3.8.1
nbformat==5.9.2
